projects/project-102-mlops-pipeline/requirements.txt
# Core dependencies
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.2

# ML tracking and experimentation
mlflow==2.8.1

# Data versioning
dvc==3.30.1
dvc-s3==2.23.0

# Cloud storage
boto3==1.29.7
s3fs==2023.12.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Kubernetes
kubernetes==28.1.0

# Monitoring
prometheus-client==0.19.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# HTTP requests
requests==2.31.0

# Job scheduling (Airflow dependencies)
apache-airflow==2.7.3
apache-airflow-providers-celery==3.4.1
apache-airflow-providers-postgres==5.7.1
apache-airflow-providers-redis==3.4.1
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
joblib==1.3.2

# Data validation
scipy==1.11.4
